Ethnic makeup of New Zealand, by region

Number of people, by ethnicity, for the New Zealand regions, based on the 2013 Estimated Resident Population.

data("region_ethnicity")

Format

A data frame with 16 observations on the following 6 variables.

Area

a factor with level Auckland region Bay of Plenty region Canterbury region Gisborne region Hawke\'s Bay region Manawatu-Wanganui region Marlborough region Nelson region Northland region Otago region Southland region Taranaki region Tasman region Waikato region Wellington region West Coast region

Asian

number of people

Euro_Other

number of people

Maori

number of people

MELAA

number of people

Pacific

number of people

Source

https://figure.nz/table/foip3RBvo69sO3J3, originally from Statistics New Zealand

Examples

data(region_ethnicity)
summary(region_ethnicity)

#>                        Area        Asian          Euro_Other    
#>  Auckland region         : 1   Min.   :   900   Min.   : 29100  
#>  Bay of Plenty region    : 1   1st Qu.:  2078   1st Qu.: 45750  
#>  Canterbury region       : 1   Median :  5725   Median :124800  
#>  Gisborne region         : 1   Mean   : 33826   Mean   :206981  
#>  Hawke's Bay region      : 1   3rd Qu.: 19750   3rd Qu.:243375  
#>  Manawatu-Wanganui region: 1   Max.   :348900   Max.   :886400  
#>  (Other)                 :10                                    
#>      Maori            MELAA          Pacific      
#>  Min.   :  3600   Min.   :  140   Min.   :   410  
#>  1st Qu.: 10888   1st Qu.:  265   1st Qu.:  1760  
#>  Median : 31250   Median :  740   Median :  5170  
#>  Mean   : 43234   Mean   : 3322   Mean   : 21529  
#>  3rd Qu.: 57850   3rd Qu.: 2750   3rd Qu.: 10558  
#>  Max.   :169800   Max.   :28200   Max.   :227000  
#>