Allocate triangles within hexagons

This implements the Webster/Sainte-Laguë method to allocate six triangles in each hexagon, in proportion to counts in two or more classes.

tri_alloc(countmatrix, colours, classes=colours, names = rownames(countmatrix))

Arguments

countmatrix
A matrix of counts with a column for each class and a row for each hexagon.

colours,classes
A vector of class names, or colour names with one for each class

names
A vector of names, for each row of countmatrix, to be applied to the output

Value

Matrix of class names, with six columns. The rownames will be set to names if provided.

Examples

data(dhb_cars)
tri_alloc(dhb_cars[,-1], c("green","gold","orange","goldenrod"), names=dhb_cars$dhb )

#>                    [,1]    [,2]   [,3]     [,4]     [,5]     [,6]       
#> Northland          "gold"  "gold" "gold"   "orange" "orange" "goldenrod"
#> Waitemata          "gold"  "gold" "orange" "orange" "orange" "goldenrod"
#> Auckland           "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Counties Manukau   "gold"  "gold" "orange" "orange" "orange" "goldenrod"
#> Waikato            "gold"  "gold" "orange" "orange" "orange" "goldenrod"
#> Lakes              "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Bay of Plenty      "gold"  "gold" "orange" "orange" "orange" "goldenrod"
#> Tairawhiti         "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Taranaki           "gold"  "gold" "gold"   "orange" "orange" "goldenrod"
#> Hawke's Bay        "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Whanganui          "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Midcentral         "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Hutt               "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Capital and Coast  "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Wairarapa          "green" "gold" "gold"   "orange" "orange" "goldenrod"
#> Nelson Marlborough "gold"  "gold" "orange" "orange" "orange" "goldenrod"
#> West Coast         "gold"  "gold" "gold"   "orange" "orange" "goldenrod"
#> Canterbury         "gold"  "gold" "orange" "orange" "orange" "goldenrod"
#> South Canterbury   "gold"  "gold" "orange" "orange" "orange" "goldenrod"
#> Southern           "green" "gold" "gold"   "orange" "orange" "goldenrod"

tri_alloc(dhb_cars[,-1], c("0","1","2","3+"), names=dhb_cars$dhb )

#>                    [,1] [,2] [,3] [,4] [,5] [,6]
#> Northland          "1"  "1"  "1"  "2"  "2"  "3+"
#> Waitemata          "1"  "1"  "2"  "2"  "2"  "3+"
#> Auckland           "0"  "1"  "1"  "2"  "2"  "3+"
#> Counties Manukau   "1"  "1"  "2"  "2"  "2"  "3+"
#> Waikato            "1"  "1"  "2"  "2"  "2"  "3+"
#> Lakes              "0"  "1"  "1"  "2"  "2"  "3+"
#> Bay of Plenty      "1"  "1"  "2"  "2"  "2"  "3+"
#> Tairawhiti         "0"  "1"  "1"  "2"  "2"  "3+"
#> Taranaki           "1"  "1"  "1"  "2"  "2"  "3+"
#> Hawke's Bay        "0"  "1"  "1"  "2"  "2"  "3+"
#> Whanganui          "0"  "1"  "1"  "2"  "2"  "3+"
#> Midcentral         "0"  "1"  "1"  "2"  "2"  "3+"
#> Hutt               "0"  "1"  "1"  "2"  "2"  "3+"
#> Capital and Coast  "0"  "1"  "1"  "2"  "2"  "3+"
#> Wairarapa          "0"  "1"  "1"  "2"  "2"  "3+"
#> Nelson Marlborough "1"  "1"  "2"  "2"  "2"  "3+"
#> West Coast         "1"  "1"  "1"  "2"  "2"  "3+"
#> Canterbury         "1"  "1"  "2"  "2"  "2"  "3+"
#> South Canterbury   "1"  "1"  "2"  "2"  "2"  "3+"
#> Southern           "0"  "1"  "1"  "2"  "2"  "3+"